- Changed sidebar tab appearance - Fixed "Home" button staying active on other pages - Made it so you cannot leave default pack even if not owned by you - Update Vite config to support JetBrains port forwarding - Fixed sidebar resize having double length when sidebar expanded

+++ apps/server/src/routes/pack/[id]/join.ts
@@ -96,7 +96,7 @@ export default (app: YapockType) =>
                     return
                 }
 
-                if (!(await PackMan.hasPermission(user.sub, id, PackMan.PERMISSIONS.Owner))) {
+                if (!(await PackMan.hasPermission(user.sub, id, PackMan.PERMISSIONS.Owner)) && user.default_pack !== id) {
                     try {
                         await prisma.packs_memberships.deleteMany({
                             where: {
